+++ webpack/common.js
@@ -13,5 +13,6 @@ module.exports = {
       extensions: ['.tsx', '.ts', '.js'],
     },
     mode: 'production',
+    devtool: 'source-map',
   },
 };